+++ checkpoint/orbax/checkpoint/experimental/v1/_src/saving/saving.py
@@ -74,6 +74,7 @@ def save(
       JSON-serializable dictionary the user can use to store additional
       information. The field is treated as opaque by Orbax.
   """
+  validation.validate_state(state)
   execution.save_checkpointables_impl(
       path,
       {checkpointable_name: state},
@@ -207,6 +208,7 @@ def save_async(
     An `AsyncResponse` that can be used to block until the save is complete.
     Blocking can be done using `response.result()`, which returns `None`.
   """
+  validation.validate_state(state)
   return execution.save_checkpointables_impl(
       path,
       {checkpointable_name: state},