+++ checkpoint/orbax/checkpoint/experimental/v1/_src/tree/types.py
@@ -21,6 +21,7 @@ import numpy as np
 from orbax.checkpoint._src.tree import types as tree_types
 from orbax.checkpoint.experimental.v1._src.arrays import types as array_types
 
+
 JsonType = tree_types.JsonType
 
 T = TypeVar("T")